+++ albam/__init__.py
@@ -55,12 +55,14 @@ def register():
     AlbamCustomPropertiesMesh = AlbamCustomPropertiesFactory("mesh")
     AlbamCustomPropertiesImage = AlbamCustomPropertiesFactory("image")
     AlbamCustomPropertiesObject = AlbamCustomPropertiesFactory("object")
+    AlbamCustomPropertiesBone = AlbamCustomPropertiesFactory("bone")
     bpy.utils.register_class(AlbamData)
     _CUSTOM_PROPERTIES_CLASSES[:] = [
         AlbamCustomPropertiesMaterial,
         AlbamCustomPropertiesMesh,
         AlbamCustomPropertiesImage,
         AlbamCustomPropertiesObject,
+        AlbamCustomPropertiesBone,
     ]
     for cls in _CUSTOM_PROPERTIES_CLASSES:
         bpy.utils.register_class(cls)
@@ -74,6 +76,8 @@ def register():
     bpy.types.Mesh.albam_custom_properties = bpy.props.PointerProperty(type=AlbamCustomPropertiesMesh)
     bpy.types.Image.albam_custom_properties = bpy.props.PointerProperty(type=AlbamCustomPropertiesImage)
     bpy.types.Object.albam_custom_properties = bpy.props.PointerProperty(type=AlbamCustomPropertiesObject)
+    # PoseBone, not Bone: see BlenderRegistry.register_custom_properties_bone
+    bpy.types.PoseBone.albam_custom_properties = bpy.props.PointerProperty(type=AlbamCustomPropertiesBone)
 
     for handler in LOAD_POST_HANDLERS:
         bpy.app.handlers.load_post.append(handler)